requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
chromadb==0.5.3
sentence-transformers==3.0.1
anthropic==0.29.0
python-dotenv==1.0.1
pydantic==2.7.4
